refactor: reorganize E2E tests
